+++ src/sql_query.ts
@@ -31,7 +31,7 @@ export default class SqlQuery {
     }
 
     replace(options, adhocFilters) {
-        var query = this.templateSrv.replace(this.target.query.trim(), options.scopedVars, SqlQuery.interpolateQueryExpr),
+        var query = this.templateSrv.replace(SqlQuery.conditionalTest(this.target.query.trim(), this.templateSrv), options.scopedVars, SqlQuery.interpolateQueryExpr),
             scanner = new Scanner(query),
             dateTimeType = this.target.dateTimeType
                 ? this.target.dateTimeType
@@ -39,6 +39,7 @@ export default class SqlQuery {
             i = this.templateSrv.replace(this.target.interval, options.scopedVars) || options.interval,
             interval = SqlQuery.convertInterval(i, this.target.intervalFactor || 1),
             adhocCondition = [];
+
         try {
             let ast = scanner.toAST();
             let topQuery = ast;
@@ -51,6 +52,7 @@ export default class SqlQuery {
                     ast.where = [];
                 }
                 let target = SqlQuery.target(ast.from[0], this.target);
+
                 adhocFilters.forEach(function (af) {
                     let parts = af.key.split('.');
                     /* Wildcard table, substitute current target table */
@@ -64,26 +66,28 @@ export default class SqlQuery {
                     /* Expect fully qualified column name at this point */
                     if (parts.length < 3) {
                         console.warn("adhoc filters: filter " + af.key + "` has wrong format");
-                        return
+                        return;
                     }
                     if (target[0] != parts[0] || target[1] != parts[1]) {
-                        return
+                        return;
                     }
                     let operator = SqlQuery.clickhouseOperator(af.operator);
-                    let cond = parts[2] + " " + operator + " " + af.value;
+                    let cond = parts[2] + " " + operator + " "
+                        + ((af.value.indexOf("'") > -1 || af.value.indexOf(", ") > -1 || af.value.match(/^\s*\d+\s*$/)) ? af.value : "'" + af.value + "'");
                     adhocCondition.push(cond);
                     if (ast.where.length > 0) {
                         // OR is not implemented
                         // @see https://github.com/grafana/grafana/issues/10918
-                        cond = "AND " + cond
+                        cond = "AND " + cond;
                     }
-                    ast.where.push(cond)
+                    ast.where.push(cond);
                 });
                 query = scanner.Print(topQuery);
             }
-            query = SqlQuery.applyMacros(query, ast)
+
+            query = SqlQuery.applyMacros(query, ast);
         } catch (err) {
-            console.error('AST parser error: ', err)
+            console.error('AST parser error: ', err);
         }
 
         /* Render the ad-hoc condition or evaluate to an always true condition */
@@ -92,6 +96,7 @@ export default class SqlQuery {
             renderedAdHocCondition = '(' + adhocCondition.join(' AND ') + ')';
         }
 
+        query = scanner.removeComments(query);
         query = SqlQuery.unescape(query);
 
         const timeRangeRaw = this.options && this.options.rangeRaw || options.rangeRaw;
@@ -101,9 +106,15 @@ export default class SqlQuery {
         }
 
         let rawQuery = query
+        let table = SqlQuery.escapeIdentifier(this.target.table);
+        if (this.target.database) {
+            table = SqlQuery.escapeIdentifier(this.target.database) + '.' + table;
+        }
+
+        this.target.rawQuery = query
             .replace(/\$timeSeries/g, SqlQuery.getTimeSeries(dateTimeType))
             .replace(/\$timeFilter/g, timeFilter)
-            .replace(/\$table/g, SqlQuery.escapeIdentifier(this.target.database) + '.' + SqlQuery.escapeIdentifier(this.target.table))
+            .replace(/\$table/g, table)
             .replace(/\$dateCol/g, SqlQuery.escapeIdentifier(this.target.dateColDataType))
             .replace(/\$dateTimeCol/g, SqlQuery.escapeIdentifier(this.target.dateTimeColDataType))
             .replace(/\$interval/g, interval)
@@ -111,25 +122,25 @@ export default class SqlQuery {
             .replace(/(?:\r\n|\r|\n)/g, ' ');
 
         const round = this.target.round === "$step"
-          ? interval
-          : SqlQuery.convertInterval(this.target.round, 1);
+            ? interval
+            : SqlQuery.convertInterval(this.target.round, 1);
         const timeRange = this.options && this.options.range || options.range;
+        this.target.rawQuery = SqlQuery.replaceTimeFilters(this.target.rawQuery, timeRange, dateTimeType, round);
 
-        rawQuery = SqlQuery.replaceTimeFilters(rawQuery, timeRange, dateTimeType, round);
-        this.target.rawQuery = rawQuery;
-
-        return rawQuery;
+        return this.target.rawQuery;
     }
 
     static escapeIdentifier(identifier: string): string {
-        if (/^[a-zA-Z_][0-9a-zA-Z_]*$/.test(identifier)) {
+        if (/^[a-zA-Z_][0-9a-zA-Z_]*$/.test(identifier)
+            || /\(.*\)/.test(identifier)
+        ) {
             return identifier;
         } else {
             return '`' + identifier.replace(/`/g, '``') + '`';
         }
     }
 
-    static replaceTimeFilters(query: string, range: TimeRange, dateTimeType : string = 'DATETIME', round?: number): string {
+    static replaceTimeFilters(query: string, range: TimeRange, dateTimeType: string = 'DATETIME', round?: number): string {
         let from = SqlQuery.convertTimestamp(SqlQuery.round(range.from, round || 0));
         let to = SqlQuery.convertTimestamp(SqlQuery.round(range.to, round || 0));
 
@@ -137,18 +148,18 @@ export default class SqlQuery {
         // data is not affected by round
         if (round > 0) {
             to += (round * 2) - 1;
-            from -= (round * 2) - 1
+            from -= (round * 2) - 1;
         }
 
         return query
-          .replace(
-            /\$timeFilterByColumn\(([\w_]+)\)/g,
-            (match: string, columnName: string) => (
-              `${ columnName } ${ SqlQuery.getFilterSqlForDateTime(range.raw.to === 'now', dateTimeType) }`
+            .replace(
+                /\$timeFilterByColumn\(([\w_]+)\)/g,
+                (match: string, columnName: string) => (
+                    `${columnName} ${SqlQuery.getFilterSqlForDateTime(range.raw.to === 'now', dateTimeType)}`
+                )
             )
-          )
-          .replace(/\$from/g, from.toString())
-          .replace(/\$to/g, to.toString());
+            .replace(/\$from/g, from.toString())
+            .replace(/\$to/g, to.toString());
     }
 
     static getFilterSqlForDateTime(isToNow: boolean, dateTimeType: string) {
@@ -164,7 +175,7 @@ export default class SqlQuery {
     static getConvertFn(dateTimeType: string) {
         return function (t: string): string {
             if (dateTimeType === 'DATETIME') {
-                return 'toDateTime('+ t +')';
+                return 'toDateTime(' + t + ')';
             }
 
             return t;
@@ -173,7 +184,7 @@ export default class SqlQuery {
 
     static target(from: string, target: any): [string, string] {
         if (from.length == 0) {
-            return ['', '']
+            return ['', ''];
         }
         let targetTable, targetDatabase;
         let parts = from.split('.');
@@ -193,7 +204,7 @@ export default class SqlQuery {
         if (targetTable === '$table') {
             targetTable = target.table;
         }
-        return [targetDatabase, targetTable]
+        return [targetDatabase, targetTable];
     }
 
     static applyMacros(query: string, ast: any): string {
@@ -212,17 +223,17 @@ export default class SqlQuery {
         if (SqlQuery.contain(ast, '$perSecondColumns')) {
             return SqlQuery.perSecondColumns(query, ast);
         }
-        return query
+        return query;
     }
 
     static contain(obj: any, field: string): boolean {
-        return obj.hasOwnProperty(field) && !isEmpty(obj[field])
+        return obj.hasOwnProperty(field) && !isEmpty(obj[field]);
     }
 
     static _parseMacros(macros: string, query: string): string {
         let mLen = macros.length;
         if (query.slice(0, mLen + 1) !== macros + '(') {
-            return ""
+            return "";
         }
         let fromIndex = SqlQuery._fromIndex(query);
         return query.slice(fromIndex);
@@ -231,7 +242,7 @@ export default class SqlQuery {
     static columns(query: string, ast: any): string {
         let q = SqlQuery._parseMacros('$columns', query);
         if (q.length < 1) {
-            return query
+            return query;
         }
         let args = ast['$columns'];
         if (args.length !== 2) {
@@ -275,7 +286,7 @@ export default class SqlQuery {
     static rateColumns(query: string, ast: any): string {
         let q = SqlQuery._parseMacros('$rateColumns', query);
         if (q.length < 1) {
-            return query
+            return query;
         }
         let args = ast['$rateColumns'];
         if (args.length !== 2) {
@@ -301,7 +312,7 @@ export default class SqlQuery {
     static rate(query: string, ast: any): string {
         let q = SqlQuery._parseMacros('$rate', query);
         if (q.length < 1) {
-            return query
+            return query;
         }
         let args = ast['$rate'];
         if (args.length < 1) {
@@ -341,7 +352,7 @@ export default class SqlQuery {
     static perSecondColumns(query: string, ast: any): string {
         let q = SqlQuery._parseMacros('$perSecondColumns', query);
         if (q.length < 1) {
-            return query
+            return query;
         }
         let args = ast['$perSecondColumns'];
         if (args.length !== 2) {
@@ -351,7 +362,15 @@ export default class SqlQuery {
         let key = args[0],
             value = 'max(' + args[1].trim() + ') AS max_0',
             havingIndex = q.toLowerCase().indexOf('having'),
-            having = "";
+            having = "",
+            aliasIndex = key.toLowerCase().indexOf(' as '),
+            alias = "perSecondColumns";
+        if (aliasIndex == -1) {
+            key = key + " AS " + alias;
+        } else {
+            alias = key.slice(aliasIndex + 4, key.length);
+        }
+
         if (havingIndex !== -1) {
             having = ' ' + q.slice(havingIndex, q.length);
             q = q.slice(0, havingIndex - 1);
@@ -360,33 +379,30 @@ export default class SqlQuery {
 
         return 'SELECT' +
             ' t,' +
-            ' groupArray((' + key + ', max_0_Rate)) AS groupArr' +
+            ' groupArray((' + alias + ', max_0_Rate)) AS groupArr' +
             ' FROM (' +
             ' SELECT t,' +
-            ' ' + key +
+            ' ' + alias +
             ', if(runningDifference(max_0) < 0, nan, runningDifference(max_0) / runningDifference(t/1000)) AS max_0_Rate' +
             ' FROM (' +
             ' SELECT $timeSeries AS t' +
             ', ' + key +
             ', ' + value + ' ' +
             q +
-            ' GROUP BY t, ' + key +
+            ' GROUP BY t, ' + alias +
             having +
-            ' ORDER BY ' + key + ', t' +
+            ' ORDER BY ' + alias + ', t' +
             ')' +
             ')' +
             ' GROUP BY t' +
             ' ORDER BY t';
-
-
-        return SqlQuery._perSecond(args, q);
     }
 
     // $perSecond(query)
     static perSecond(query: string, ast: any): string {
         let q = SqlQuery._parseMacros('$perSecond', query);
         if (q.length < 1) {
-            return query
+            return query;
         }
         let args = ast['$perSecond'];
         if (args.length < 1) {
@@ -394,7 +410,7 @@ export default class SqlQuery {
         }
 
         each(args, function (a, i) {
-            args[i] = 'max(' + a.trim() + ') AS max_' + i
+            args[i] = 'max(' + a.trim() + ') AS max_' + i;
         });
 
         return SqlQuery._perSecond(args, q);
@@ -434,14 +450,14 @@ export default class SqlQuery {
         if (dateTimeType === 'DATETIME') {
             return '(intDiv(toUInt32($dateTimeCol), $interval) * $interval) * 1000';
         }
-        return '(intDiv($dateTimeCol, $interval) * $interval) * 1000'
+        return '(intDiv($dateTimeCol, $interval) * $interval) * 1000';
     }
 
     static getDateFilter(isToNow: boolean) {
         if (isToNow) {
             return '$dateCol >= toDate($from)';
         }
-        return '$dateCol BETWEEN toDate($from) AND toDate($to)'
+        return '$dateCol BETWEEN toDate($from) AND toDate($to)';
     }
 
     static getDateTimeFilter(isToNow: boolean, dateTimeType: string) {
@@ -449,7 +465,7 @@ export default class SqlQuery {
             if (dateTimeType === 'DATETIME') {
                 return 'toDateTime(' + t + ')';
             }
-            return t
+            return t;
         };
 
         if (isToNow) {
@@ -484,7 +500,7 @@ export default class SqlQuery {
 
     static convertInterval(interval: any, intervalFactor: number): number {
         if (interval === undefined || typeof interval !== 'string' || interval == "") {
-            return 0
+            return 0;
         }
         var m = interval.match(durationSplitRegexp);
         if (m === null) {
@@ -527,7 +543,7 @@ export default class SqlQuery {
                 return "NOT LIKE";
             default:
                 console.warn("adhoc filters: got unsupported operator `" + value + "`");
-                return value
+                return value;
         }
     }
 
@@ -542,7 +558,7 @@ export default class SqlQuery {
                 isDigit = false;
                 return false;
             }
-            return true
+            return true;
         });
 
         if (isDigit) {
@@ -552,11 +568,60 @@ export default class SqlQuery {
         }
     }
 
+    static conditionalTest(query, templateSrv) {
+        let macros = '$conditionalTest(';
+        let openMacros = query.indexOf(macros);
+        while (openMacros !== -1) {
+            let r = SqlQuery.betweenBraces(query.substring(openMacros + macros.length, query.length));
+            if (r.error.length > 0) {
+                throw {message: '$conditionalIn macros error: ' + r.error};
+            }
+            let arg = r.result;
+            // first parameters is an expression and require some complex parsing,
+            // so parse from the end where you know that the last parameters is a comma with a variable
+            let param1 = arg.substring(0, arg.lastIndexOf(',')).trim();
+            let param2 = arg.substring(arg.lastIndexOf(',') + 1).trim();
+            // remove the $ from the variable
+            let varinparam = param2.substring(1);
+            let done = 0;
+            //now find in the list of variable what is the value
+            for (var i = 0; i < templateSrv.variables.length; i++) {
+                var varG = templateSrv.variables[i];
+                if (varG.name === varinparam) {
+                    let closeMacros = openMacros + macros.length + r.result.length + 1;
+                    done = 1;
+
+                    const value = varG.current.value;
+
+                    if (
+                        // for query variable when all is selected
+                        // may be add another test on the all activation may be wise.
+                        (varG.type === 'query' && ((value.length == 1 && value[0] === '$__all')
+                            || (typeof value === 'string' && value === '$__all'))) ||
+                        // for textbox variable when nothing is entered
+                        (['textbox', 'custom'].includes(varG.type) && ['', undefined, null].includes(value))) {
+                        query = query.substring(0, openMacros) + ' ' + query.substring(closeMacros, query.length);
+                    } else {
+                        // replace of the macro with standard test.
+                        query = query.substring(0, openMacros) + ' ' + param1 + ' ' + query.substring(closeMacros, query.length);
+                    }
+                    break;
+                }
+            }
+            if (done == 0) {
+                throw {message: '$conditionalTest macros error cannot find referenced variable: ' + param2};
+            }
+            openMacros = query.indexOf(macros);
+        }
+        return query;
+    }
+
+
     static unescape(query) {
         let macros = '$unescape(';
         let openMacros = query.indexOf(macros);
         while (openMacros !== -1) {
-            let r = SqlQuery.betweenBraces(query.substring(openMacros+macros.length, query.length));
+            let r = SqlQuery.betweenBraces(query.substring(openMacros + macros.length, query.length));
             if (r.error.length > 0) {
                 throw {message: '$unescape macros error: ' + r.error};
             }
@@ -566,7 +631,7 @@ export default class SqlQuery {
             query = query.substring(0, openMacros) + arg + query.substring(closeMacros, query.length);
             openMacros = query.indexOf('$unescape(');
         }
-        return query
+        return query;
     }
 
     static betweenBraces(query): any {
@@ -588,8 +653,8 @@ export default class SqlQuery {
             }
         }
         if (openBraces > 1) {
-            r.error = "missing parentheses"
+            r.error = "missing parentheses";
         }
-        return r
+        return r;
     }
 }